requirements.txt
# local package
-e .

# external requirements
alabaster=0.7.12=pypi_0
anyio=3.3.0=pypi_0
argon2-cffi=20.1.0=py39h27cfd23_1
async_generator=1.10=pyhd3eb1b0_0
attrs=21.2.0=pyhd3eb1b0_0
autopep8=1.5.7=pypi_0
awscli=1.20.11=pypi_0
babel=2.9.1=pyh44b312d_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
bleach=3.3.1=pypi_0
botocore=1.21.11=pypi_0
bottleneck=1.3.2=py39hdd57654_1
brotli=1.0.9=he6710b0_2
brotlipy=0.7.0=py39h3811e60_1001
ca-certificates=2021.10.8=ha878542_0
certifi=2021.10.8=py39hf3d152e_1
cffi=1.14.6=py39h400218f_0
charset-normalizer=2.0.9=pyhd8ed1ab_0
click=8.0.1=pypi_0
colorama=0.4.3=pypi_0
coverage=5.5=pypi_0
cryptography=35.0.0=py39hbca0aa6_0
cycler=0.10.0=pypi_0
dbus=1.13.18=hb2f20db_0
debugpy=1.5.1=py39h295c915_0
decorator=5.1.0=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
docutils=0.15.2=pypi_0
entrypoints=0.3=pypi_0
expat=2.4.1=h2531618_2
flake8=3.9.2=pypi_0
flake8-isort=4.0.0=pypi_0
fontconfig=2.13.1=h6c09931_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.11.0=h70c0345_0
giflib=5.2.1=h7b6447c_0
glib=2.69.1=h5202010_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
h11=0.12.0=pyhd3eb1b0_0
h2=4.0.0=py39h06a4308_3
hpack=4.0.0=py_0
httpcore=0.12.3=pyhd3eb1b0_0
httpx=0.17.1=pyhd3eb1b0_0
hyperframe=6.0.1=pyhd3eb1b0_0
icu=58.2=he6710b0_3
idna=3.2=pypi_0
imagesize=1.2.0=pypi_0
importlib-metadata=4.8.2=py39h06a4308_0
importlib_metadata=4.8.2=hd3eb1b0_0
intel-openmp=2021.4.0=h06a4308_3561
ipdb=0.13.9=pypi_0
ipykernel=6.0.3=pypi_0
ipython=7.29.0=py39hb070fc8_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.5=pyhd3eb1b0_1
isort=5.9.3=pypi_0
jedi=0.18.0=py39h06a4308_1
jinja2=3.0.2=pyhd3eb1b0_0
jmespath=0.10.0=pypi_0
joblib=1.1.0=pyhd3eb1b0_0
jpeg=9d=h7f8727e_0
json5=0.9.5=pyh9f0ad1d_0
jsonschema=3.2.0=pyhd3eb1b0_2
jupyter=1.0.0=pypi_0
jupyter-console=6.4.0=pypi_0
jupyter-server=1.10.1=pypi_0
jupyter_client=7.0.6=pyhd3eb1b0_0
jupyter_core=4.9.1=py39h06a4308_0
jupyter_server=1.13.1=pyhd8ed1ab_0
jupyterlab=3.1.1=pypi_0
jupyterlab-server=2.6.1=pypi_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_server=2.10.1=pyhd8ed1ab_0
jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
kiwisolver=1.3.1=py39h2531618_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libgomp=9.3.0=h5101ec6_17
libpng=1.6.37=hbc83047_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libtiff=4.2.0=h85742a9_0
libuuid=1.0.3=h7f8727e_2
libwebp=1.2.0=h89dd481_0
libwebp-base=1.2.0=h27cfd23_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.12=h03d6c58_0
libxslt=1.1.34=hc22bd24_0
lxml=4.7.1=py39h1f438cf_1
lz4-c=1.9.3=h295c915_1
markupsafe=2.0.1=py39h27cfd23_0
matplotlib=3.5.0=py39h06a4308_0
matplotlib-base=3.5.0=py39h3ed280b_0
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mccabe=0.6.1=pypi_0
mistune=0.8.4=py39h27cfd23_1000
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py39h7f8727e_0
mkl_fft=1.3.1=py39hd3c417c_0
mkl_random=1.2.2=py39h51133e4_0
munkres=1.1.4=py_0
nbclassic=0.3.1=pypi_0
nbclient=0.5.3=pyhd3eb1b0_0
nbconvert=6.1.0=py39h06a4308_0
nbformat=5.1.3=pyhd3eb1b0_0
ncurses=6.3=h7f8727e_2
nest-asyncio=1.5.1=pyhd3eb1b0_0
notebook=6.4.0=pypi_0
numexpr=2.7.3=py39h22e1b3c_1
numpy=1.21.1=pypi_0
numpy-base=1.21.2=py39h79a1101_0
olefile=0.46=pyhd3eb1b0_0
openssl=1.1.1l=h7f8727e_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.3.1=pypi_0
pandocfilters=1.4.3=py39h06a4308_1
parso=0.8.2=pyhd3eb1b0_0
patsy=0.5.2=py39h06a4308_0
pcre=8.45=h295c915_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.4.0=py39h5aabda8_0
pip=21.2.2=pypi_0
plotly=5.1.0=pyhd3eb1b0_0
prometheus_client=0.12.0=pyhd3eb1b0_0
prompt-toolkit=3.0.19=pypi_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pyasn1=0.4.8=pypi_0
pycodestyle=2.7.0=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pyflakes=2.3.1=pypi_0
pygments=2.9.0=pypi_0
pyopenssl=21.0.0=pyhd8ed1ab_0
pyparsing=3.0.4=pyhd3eb1b0_0
pyqt=5.9.2=py39h2531618_6
pyrsistent=0.18.0=py39heee7806_0
pysocks=1.7.1=py39hf3d152e_4
python=3.9.7=h12debd9_1
python-dateutil=2.8.2=pyhd3eb1b0_0
python-dotenv=0.19.0=pypi_0
python_abi=3.9=2_cp39
pytz=2021.3=pyhd3eb1b0_0
pyyaml=5.4.1=pypi_0
pyzmq=22.1.0=pypi_0
qt=5.9.7=h5867ecd_1
qtconsole=5.1.1=pypi_0
qtpy=1.9.0=pypi_0
readline=8.1=h27cfd23_0
requests=2.26.0=pyhd8ed1ab_1
requests-unixsocket=0.2.0=pypi_0
rfc3986=1.4.0=pyhd3eb1b0_0
rsa=4.7.2=pypi_0
rstcheck=3.3.1=pypi_0
s3transfer=0.5.0=pypi_0
scikit-learn=1.0.1=py39h51133e4_0
scipy=1.7.1=py39h292c36d_2
seaborn=0.11.2=pyhd3eb1b0_0
send2trash=1.7.1=pypi_0
setuptools=57.4.0=pypi_0
sip=4.19.13=py39h2531618_0
six=1.15.0=py39h06a4308_0
sniffio=1.2.0=py39h06a4308_1
snowballstemmer=2.1.0=pypi_0
sphinx=4.1.2=pypi_0
sphinxcontrib-applehelp=1.0.2=pypi_0
sphinxcontrib-devhelp=1.0.2=pypi_0
sphinxcontrib-htmlhelp=2.0.0=pypi_0
sphinxcontrib-jsmath=1.0.1=pypi_0
sphinxcontrib-qthelp=1.0.3=pypi_0
sphinxcontrib-serializinghtml=1.1.5=pypi_0
sqlite=3.36.0=hc218d9a_0
src=0.1.0=dev_0
statsmodels=0.13.0=py39h7f8727e_0
tenacity=8.0.1=py39h06a4308_0
terminado=0.10.1=pypi_0
testfixtures=6.18.0=pypi_0
testpath=0.5.0=pyhd3eb1b0_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.11=h1ccaba5_0
toml=0.10.2=pypi_0
tornado=6.1=py39h27cfd23_0
tqdm=4.62.0=pypi_0
traitlets=5.1.1=pyhd3eb1b0_0
tzdata=2021e=hda174b7_0
urllib3=1.26.7=pyhd8ed1ab_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=pypi_0
websocket-client=1.1.0=pypi_0
wheel=0.37.0=pyhd3eb1b0_1
widgetsnbextension=3.5.1=py39h06a4308_0
xlrd=2.0.1=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
zeromq=4.3.4=h2531618_0
zipp=3.6.0=pyhd3eb1b0_0
zlib=1.2.11=h7f8727e_4
zstd=1.4.9=haebb681_0
